fixed matcher endpoint and pep8 issues
