+++ app/routers/mentor_router.py
@@ -13,17 +13,17 @@ Router = APIRouter(
 Router.db = MongoDB()
 
 
-@Router.post("/read/mentor")
-async def read_mentor(data: Optional[Dict] = None):
-    return {"result": Router.db.read("Mentors", data)}
-
-
 @Router.post("/create/mentor")
 async def create_mentor(data: Mentor):
     try:
         return {"result": Router.db.create("Mentors", data.dict())}
     except DuplicateKeyError:
         raise HTTPException(status_code=409, detail="Profile ID must be unique.")
+
+
+@Router.post("/read/mentor")
+async def read_mentor(data: Optional[Dict] = None):
+    return {"result": Router.db.read("Mentors", data)}
 
 
 @Router.post("/update/mentor/{profile_id}")
