+++ app/routers/model_router.py
@@ -1,19 +1,13 @@
 from fastapi import APIRouter
+
 from app.data import MongoDB
-from app.schema import Feedback
-from app.sentiment import sentiment_rank
+from app.model import MatcherSortSearch
 
 Router = APIRouter(
     tags=["Model Operations"],
 )
 Router.db = MongoDB()
-
-
-@Router.post("/create/feedback")
-async def create_feedback(data: Feedback):
-    data_dict = data.dict(exclude_none=True)
-    data_dict["vader_score"] = sentiment_rank(data_dict["text"])
-    return {"result": Router.db.create("Feedback", data_dict)}
+Router.matcher = MatcherSortSearch()
 
 
 @Router.post("/match/{profile_id}")
